# app/routes/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Content-X (Groot) API",
    "description": "Schema-driven CMS API for the Fynd Platform extension **groot**. Routes under `/api` are served through the FDK extension router (`fdk-extension-javascript`); clients must have a valid extension / platform session (cookies and OAuth flow). `company_id` and `application_id` are taken from session, query, or body per `request-context.helper.js`. Published CMS instance routes use `router.param('slug', loadDefinition)` and require a **published** definition for that slug.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as the extension (paths include `/api/v1/...`)"
    }
  ],
...
  "paths": {
...
    "/api/v1/cms/{slug}/bulk/import-async": {
      "post": {
        "tags": ["CMS instances"],
        "summary": "Start async bulk import with spreadsheet and assets ZIP",
        "operationId": "startBulkImportAsync",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["spreadsheet", "assets"],
                "properties": {
                  "spreadsheet": { "type": "string", "format": "binary", "description": "CSV or XLSX data file" },
                  "assets": { "type": "string", "format": "binary", "description": "ZIP of asset files referenced by basename in spreadsheet" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkImportJobStarted" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/DefinitionNotFound" }
        }
      }
    },
    "/api/v1/cms/{slug}/bulk/jobs": {
      "get": {
        "tags": ["CMS instances"],
        "summary": "List recent bulk import jobs",
        "operationId": "listBulkImportJobs",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/BulkImportJobStatus" }
                    }
                  }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/ForbiddenEmbeddedOnly" },
          "404": { "$ref": "#/components/responses/DefinitionNotFound" }
        }
      }
    },
    "/api/v1/cms/{slug}/bulk/jobs/{jobId}": {
      "get": {
        "tags": ["CMS instances"],
        "summary": "Get bulk import job status",
        "operationId": "getBulkImportJobStatus",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkImportJobStatus" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v1/cms/{slug}/bulk/jobs/{jobId}/report": {
      "get": {
        "tags": ["CMS instances"],
        "summary": "Download bulk import result report",
        "operationId": "downloadBulkImportReport",
        "parameters": [
          { "$ref": "#/components/parameters/Slug" },
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "format",
            "in": "query",
            "schema": { "type": "string", "enum": ["csv", "xlsx"] }
          },
          { "name": "application_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Report file download",
            "content": {
              "text/csv": { "schema": { "type": "string", "format": "binary" } },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "Slug": {
        "name": "slug",
        "in": "path",
        "required": true,
        "description": "Published object definition slug (loads definition via `loadDefinition`).",
        "schema": { "type": "string" }
      }
    },
    "schemas": {
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "BulkImportJobStarted": {
        "type": "object",
        "properties": {
          "job_id": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "processing", "completed", "failed"] }
        }
      },
      "BulkImportJobStatus": {
        "type": "object",
        "properties": {
          "job_id": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "processing", "completed", "failed"] },
          "total_rows": { "type": "integer" },
          "processed_rows": { "type": "integer" },
          "created_count": { "type": "integer" },
          "failed_count": { "type": "integer" },
          "asset_count": { "type": "integer" },
          "error_message": { "type": "string" },
          "can_download_report": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "started_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time" }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
...
      "DefinitionNotFound": {
        "description": "Published object definition not found for slug",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      },
      "ForbiddenEmbeddedOnly": {
        "description": "Definition is embedded_only; standalone instance APIs are disabled",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorMessage" }
          }
        }
      }
    }
  }
}